[Feat/#304] :wrench: Refactor: 이너 클래스로 이동
