--- chatty-be/src/main/java/com/chatty/chatty/game/controller/dto/PlayerScoreDTO.java
@@ -1,11 +1,0 @@
-package com.chatty.chatty.game.controller.dto;
-
-import lombok.Builder;
-
-@Builder
-public record PlayerScoreDTO(
-        Long playerId,
-        String nickname,
-        Integer score
-) {
-}
