+++ chatty-be/src/main/java/com/chatty/chatty/game/controller/dto/ScoreResponse.java
@@ -7,4 +7,11 @@ import lombok.Builder;
 public record ScoreResponse(
         List<PlayerScoreDTO> scores
 ) {
+    @Builder
+    public record PlayerScoreDTO(
+            Long playerId,
+            String nickname,
+            Integer score
+    ) {
+    }
 }
